+++ tests/test_utils.py
@@ -1,8 +1,5 @@
 """Testing the module for utils in mock-alchemy."""
 import pytest
-from sqlalchemy import Column
-from sqlalchemy import Integer
-from sqlalchemy import String
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm.exc import MultipleResultsFound
 
@@ -14,6 +11,8 @@ from mock_alchemy.utils import indexof
 from mock_alchemy.utils import match_type
 from mock_alchemy.utils import raiser
 from mock_alchemy.utils import setattr_tmp
+
+from .common import SomeClass
 
 Base = declarative_base()
 
@@ -72,16 +71,6 @@ def test_raiser() -> None:
 
 def test_idmap() -> None:
     """Tests building an idmap."""
-
-    class SomeClass(Base):
-        __tablename__ = "some_table"
-        pk1 = Column(Integer, primary_key=True)
-        pk2 = Column(Integer, primary_key=True)
-        name = Column(String(50))
-
-        def __repr__(self) -> str:
-            return str(self.pk1)
-
     expected_idmap = {(1, 2): 1}
     idmap = build_identity_map([SomeClass(pk1=1, pk2=2)])
     assert str(expected_idmap) == str(idmap)
@@ -97,16 +86,9 @@ def test_get_attr() -> None:
 
 def test_get_scalar() -> None:
     """Tests utility for getting scalar values."""
-
-    class GetScalar(Base):
-        __tablename__ = "get_scalar"
-        pk1 = Column(Integer, primary_key=True)
-        pk2 = Column(Integer, primary_key=True)
-        name = Column(String(50))
-
-    assert 1 == get_scalar([GetScalar(pk1=1, pk2=2)])
-    assert None is get_scalar([GetScalar(pk1=None, pk2=2)])
+    assert 1 == get_scalar([SomeClass(pk1=1, pk2=2)])
+    assert None is get_scalar([SomeClass(pk1=None, pk2=2)])
     assert None is get_scalar([])
     with pytest.raises(MultipleResultsFound):
-        get_scalar([GetScalar(pk1=1, pk2=2), GetScalar(pk1=2, pk2=3)])
-    assert None is get_scalar([GetScalar(pk1=None, pk2=None)])
+        get_scalar([SomeClass(pk1=1, pk2=2), SomeClass(pk1=2, pk2=3)])
+    assert None is get_scalar([SomeClass(pk1=None, pk2=None)])
